requirements.txt
# Core ML/Data Science
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.17.0

# MLOps Tools
dvc==3.48.0
clearml==1.15.0
mlflow==2.9.2

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Validation
great-expectations==0.18.5

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
